Panaversity_Course/OpenAI_AgentSDK/Q4_AgenticAI_Repo/01_ai_agents_first/15_advanced_tools/pyproject.toml
[project]
name = "15-advanced-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents==0.2.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
